refactor: moved maplibre style components to under maplibre folder
