+++ sites/geohub/src/components/maplibre/MaplibreColorPicker.svelte
