+++ sites/geohub/src/components/maplibre/OpacityPanel.svelte
