+++ sites/geohub/src/components/maplibre/fill/FillOutlineColor.test.ts
