+++ sites/geohub/src/components/maplibre/heatmap/HeatmapColorRow.svelte
