+++ sites/geohub/src/components/maplibre/heatmap/Slider.svelte
