+++ sites/geohub/src/components/maplibre/raster/RasterHueRotate.svelte
@@ -1,14 +1,14 @@
 <script lang="ts">
 	import { MAPSTORE_CONTEXT_KEY, type MapStore } from '$stores';
 	import { getContext } from 'svelte';
-	import Slider from './Slider.svelte';
+	import Slider from '$components/maplibre/raster/Slider.svelte';
 
 	const map: MapStore = getContext(MAPSTORE_CONTEXT_KEY);
 
 	export let layerId: string;
 
 	const getValue = () => {
-		let value = $map.getPaintProperty(layerId, 'raster-brightness-min');
+		let value = $map.getPaintProperty(layerId, 'raster-hue-rotate');
 
 		if (!value) {
 			value = 0;
@@ -21,8 +21,8 @@
 	$: value, setValue();
 
 	const setValue = () => {
-		map.setPaintProperty(layerId, 'raster-brightness-min', value);
+		map.setPaintProperty(layerId, 'raster-hue-rotate', value);
 	};
 </script>
 
-<Slider bind:value min={0} max={1} step={0.1} unit="" />
+<Slider bind:value min={0} max={359} step={1} unit="°" />
