+++ sites/geohub/src/components/maplibre/raster/Slider.svelte
