+++ sites/geohub/src/components/maplibre/symbol/IconColor.test.ts
