+++ sites/geohub/src/components/maplibre/symbol/IconImagePicker.test.ts
