+++ sites/geohub/src/components/maplibre/symbol/IconImagePickerCard.svelte
