+++ sites/geohub/src/components/maplibre/symbol/IconOffset.svelte
