+++ sites/geohub/src/components/maplibre/symbol/IconOverlap.svelte
