+++ sites/geohub/src/components/maplibre/symbol/IconOverlap.test.ts
