+++ sites/geohub/src/components/maplibre/symbol/IconSize.svelte
