+++ sites/geohub/src/components/maplibre/symbol/PropertySelect.svelte
