+++ sites/geohub/src/components/maplibre/symbol/PropertySelect.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, beforeEach } from 'vitest';
 import { render, type RenderResult } from '@testing-library/svelte';
-import PropertySelect from '$components/controls/vector-styles/PropertySelect.svelte';
+import PropertySelect from '$components/maplibre/symbol/PropertySelect.svelte';
 
 const layer = {
 	id: 'e078c8ff-2838-4c2f-b8c0-12793355f8ee',
