+++ sites/geohub/src/components/maplibre/symbol/SymbolPlacement.svelte
