+++ sites/geohub/src/components/maplibre/symbol/TextHaloCalor.svelte
@@ -2,7 +2,7 @@
 	import type { Layer } from '$lib/types';
 	import type { LayerSpecification } from 'maplibre-gl';
 	import { createEventDispatcher, getContext } from 'svelte';
-	import MaplibreColorPicker from './MaplibreColorPicker.svelte';
+	import MaplibreColorPicker from '$components/maplibre/MaplibreColorPicker.svelte';
 	import { MAPSTORE_CONTEXT_KEY, type MapStore } from '$stores';
 
 	const map: MapStore = getContext(MAPSTORE_CONTEXT_KEY);
