+++ sites/geohub/src/components/maplibre/symbol/TextHaloWidth.svelte
