+++ sites/geohub/src/components/maplibre/symbol/TextMaxWidth.svelte
