+++ sites/geohub/src/components/pages/map/layers/vector/PropertySelectButtons.svelte
