+++ sites/geohub/src/components/pages/map/layers/vector/VectorFilter.svelte
@@ -2,7 +2,7 @@
 	import Step from '$components/util/Step.svelte';
 	import Wizard from '$components/util/Wizard.svelte';
 	import OperationButtons from '$components/pages/map/layers/vector/OperationButtons.svelte';
-	import PropertySelectButtons from '$components/controls/vector-styles/PropertySelectButtons.svelte';
+	import PropertySelectButtons from '$components/pages/map/layers/vector/PropertySelectButtons.svelte';
 	import ValueInput from '$components/pages/map/layers/vector/ValueInput.svelte';
 	import { VectorFilterOperators } from '$lib/config/AppConfig';
 	import { clean, getLayerStyle } from '$lib/helper';
